Features-A5-ModularSystemForFunc
- Начато создание модульной системы для функций и  их вызова
